// src/app/scheduling/model/calender.model.ts
export interface CalenderModule {
  appMessageId: string;
  sendToUser: string;
  messageText: string;
  scheduleMessage: string;
  sendDate: string;
  slotId: string;
  reqId: string;
  createdAt: string;
  updatedAt: string;
  createdBy: string;
  updatedBy: string;
  isDelete: string;
  status: string;
  messageTitle: string;
  readCount: number;
  sendToGroup: string;
  recordStatus: string;
  startDate: string;
  endDate: string;
  appointmentType: string;
  messageDetailsId?: string;
}